Completed the Button Functionality
The button's click event now creates a new grid while checking to see if an old one exists and removes it if it does. The maximum grid size will be 100x100, so a
conditional was also added to see if the number entered is acceptable otherwise it will default to the maximum of 100.