Added Two Enhancements To The Hover Effect
I have added a darkening effect and a random color effect to the hover event listener. Each square will have an opacity of 0.1. Now every time the mouse hovers over
a square, the background color will change randomly and the opacity will increase by 0.1 which gives it the 'darkening' effect.